cipher: refactor asynchronous stream cipher traits
